omgezet naar PlatusType
