+++ src/platus/gegevens/types.py
@@ -1,4 +1,4 @@
-from typing import Dict, Any
+from typing import Dict, List, Any
 
 
 class PlatusType:
@@ -7,7 +7,7 @@ class PlatusType:
     def van_json(
         cls,
         **dict: Dict[str, Any]
-    ):
+        ):
         
         return cls(**dict)
     
@@ -16,7 +16,6 @@ class PlatusType:
         dict    =   {}
         
         for veld, waarde in self.__dict__.items():
-            
             if veld == "type":
                 continue
             if isinstance(waarde, list):
@@ -30,3 +29,189 @@ class PlatusType:
             dict[veld]    =   waarde
         
         return dict
+
+class HoofdCategorie(PlatusType):
+    
+    def __init__(
+        self,
+        naam    :   str,
+        type    :   str,
+        ) -> "HoofdCategorie":
+        
+        self.naam   =   naam
+        self.type   =   type
+
+class Categorie(PlatusType):
+    
+    def __init__(
+        self,
+        naam           :   str,
+        hoofdcat_uuid  :   str,
+        trefwoorden    :   list  =   None,
+        ):
+        
+        self.naam           =   naam
+        self.hoofdcat_uuid  =   hoofdcat_uuid
+        self.trefwoorden    =   list() if trefwoorden is None else trefwoorden
+
+class Land(PlatusType):
+    
+    def __init__(self,
+        naam: str,
+        iso_3166_1_alpha_3: str,
+        synoniemen: List[str] = None,
+        ) -> "Land":
+        
+        self.naam               =   naam
+        self.iso_3166_1_alpha_3 =   iso_3166_1_alpha_3
+        self.synoniemen         =   list() if synoniemen is None else synoniemen
+
+class Locatie(PlatusType):
+    
+    def __init__(
+        self,
+        naam:           str,
+        land_uuid:      str,
+        breedtegraad:   float,
+        lengtegraad:    float,
+        synoniemen:     List[str] = None
+        ) -> "Locatie":
+        
+        self.naam           =   naam
+        self.land_uuid      =   land_uuid
+        self.breedtegraad   =   breedtegraad
+        self.lengtegraad    =   lengtegraad
+        self.synoniemen     =   list() if synoniemen is None else synoniemen
+
+class Derde(PlatusType):
+    
+    def __init__(
+        self, 
+        naam            :   str,
+        type            :   str,
+        iban            :   List[str],
+        rekeningnummer  :   List[str],
+        ) -> "Derde":
+        
+        self.naam           =   naam
+        self.type           =   type
+        self.iban           =   list() if iban is None else iban
+        self.rekeningnummer =   list() if rekeningnummer is None else rekeningnummer
+    
+    def toevoegen_iban(
+        self,
+        iban_nieuw:  str,
+        ) -> "Derde":
+        
+        if not iban_nieuw in self.iban:
+            self.iban.append(iban_nieuw)
+        
+        return self
+
+    def toevoegen_rekeningnummer(
+        self,
+        rekeningnummer_nieuw:  str,
+        ) -> "Derde":
+        
+        if not rekeningnummer_nieuw in self.rekeningnummer:
+            self.rekeningnummer.append(rekeningnummer_nieuw)
+        
+        return self
+    
+    def toevoegen_giro(
+        self,
+        giro_nieuw:  str,
+        ) -> "Derde":
+        
+        if not giro_nieuw in self.giro:
+            self.giro.append(giro_nieuw)
+        
+        return self
+    
+class Persoon(Derde):
+    
+    def __init__(
+        self,
+        naam           :   str,
+        groep          :   str         =   "ongegroepeerd",
+        rekeningnummer :   List[str]   =   None,
+        iban           :   List[str]   =   None,
+        giro           :   List[str]   =   None,
+        ) -> "Persoon":
+        
+        super().__init__(
+            naam           =   naam,
+            type           =   "persoon",
+            iban           =   iban,
+            rekeningnummer =   rekeningnummer,
+            )
+        
+        self.groep  =   groep
+        self.giro   =   list() if giro is None else giro
+
+class Bedrijf(Derde):
+
+    def __init__(
+        self,
+        naam           :   str,
+        synoniemen     :   list        =   None,
+        uitsluiten     :   bool        =   False,
+        rekeningnummer :   List[str]   =   None,
+        iban           :   List[str]   =   None,
+        giro           :   List[str]   =   None,
+        cat_uuid       :   str         =   None,
+        ) -> "Bedrijf":
+        
+        super().__init__(
+            naam           =   naam,
+            type           =   "bedrijf",
+            iban           =   iban,
+            rekeningnummer =   rekeningnummer,
+            )
+        
+        self.synoniemen =   list() if synoniemen is None else synoniemen
+        self.giro       =   list() if giro is None else giro
+        self.uitsluiten =   uitsluiten
+        self.cat_uuid   =   cat_uuid
+
+class Cpsp(Derde):
+
+    def __init__(
+        self,
+        naam           :   str,
+        synoniemen     :   List[str]   =   None,
+        uitsluiten     :   bool        =   False,
+        rekeningnummer :   List[str]   =   None,
+        iban           :   List[str]   =   None,
+        ) -> "Cpsp":
+        
+        super().__init__(
+            naam           =   naam,
+            type           =   "cpsp",
+            iban           =   iban,
+            rekeningnummer =   rekeningnummer,
+            )
+        
+        self.synoniemen =   list() if synoniemen is None else synoniemen
+        self.uitsluiten =   uitsluiten
+
+class Bank(Derde):
+
+    def __init__(
+        self,
+        naam           :   str,
+        synoniemen     :   List[str]   =   None,
+        rekeningnummer :   List[str]   =   None,
+        iban           :   List[str]   =   None,
+        bic            :   List[str]   =   None,
+        ) -> "Bank":
+        
+        super().__init__(
+            naam           =   naam,
+            type           =   "bank",
+            iban           =   iban,
+            rekeningnummer =   rekeningnummer,
+            )
+        
+        self.synoniemen =   list() if synoniemen is None else synoniemen
+        self.bic        =   list() if bic is None else bic
